organize club service file
