+++ src/service/nestjs/src/api/api.module.ts
@@ -3,9 +3,8 @@ import { AuthModule } from './auth/auth.module';
 import { ClubsModule } from './club/clubs.module';
 import { UserModule } from './user/user.module';
 import { receiptModule } from './receipt/receipt.module';
-import { ReviewsModule } from './review/reviews.module';
 
 @Module({
-	imports: [AuthModule, ClubsModule, UserModule, receiptModule, ReviewsModule],
+	imports: [AuthModule, ClubsModule, UserModule, receiptModule],
 })
 export class ApiModule {}
