+++ src/service/nestjs/src/api/club/clubs.controller.ts
@@ -31,6 +31,8 @@ import { Types } from 'mongoose';
 import { ScheduleService } from './service';
 import { Response } from 'express';
 import * as ClubDto from './dto/index';
+import { ReviewsService } from './service/reviews.service';
+import { CreateReviewDTO } from './dto/request/createReview.dto';
 
 @ApiTags('club API')
 @Controller('clubs')
@@ -38,6 +40,7 @@ export class ClubsController {
 	constructor(
 		private readonly clubsService: ClubsService,
 		private readonly scheduleService: ScheduleService,
+		private readonly reveiwSerice: ReviewsService,
 	) {}
 
 	@Get()
@@ -183,6 +186,41 @@ export class ClubsController {
 		} catch (error) {
 			console.error(error);
 			throw new HttpException(error.message, error.status);
+		}
+	}
+
+	// @Post(':cid/:uid')
+	@Post()
+	@ApiOperation({ summary: 'create a club review', description: 'review 생성' })
+	@ApiBody({ type: CreateReviewDTO })
+	// @ApiParam({ name: 'cid', description: 'Club ID' })
+	// @ApiParam({ name: 'uid', description: 'User ID' })
+	@ApiResponse({ status: 201, description: '업로드에 성공하였습니다' })
+	@ApiResponse({ status: 404, description: '업로드에 실패하였습니다' })
+	async create(
+		@Body() createReivewDTO: CreateReviewDTO,
+		// 그냥 Body로 id를 받아오면 안되나?
+		// @Param('cid') clubId: string,
+		// @Param('uid') userId: string,
+	) {
+		try {
+			const promiseReview = this.reveiwSerice.create(createReivewDTO);
+			if (!promiseReview) {
+				throw new NotFoundException('review not found');
+			}
+
+			// //club의 reviews배열에 reviewId 추가
+			// const clubId = (await promiseReview.then()).clubId; // 리뷰의 clubID 추출
+			// const club = await this.clubsService.findOne(clubId); // clubId로 club을 찾음
+			// if (!club) {
+			// 	throw new NotFoundException('review not found');
+			// }
+			// club.reviews.push((await promiseReview.then())._id); // club의 reviews배열에 reviewId 추가
+			// await this.clubsService.update(clubId, club); // club 업데이트
+
+			return promiseReview;
+		} catch (error) {
+			throw new HttpException('업로드에 실패하였습니다', HttpStatus.BAD_REQUEST);
 		}
 	}
 }
